Addicion de Escritura y Lectura de Ficheros Binarios
Falta Organizar la Escritura para que la lectura sea mas amigable al usuario

+++ src/sample/Edit.java
@@ -259,7 +259,7 @@ public class Edit {
     public void changetenPuls(TableColumn.CellEditEvent edittedCell)
     {
         EHipertension personselected = (EHipertension) tblvwTension.getSelectionModel().getSelectedItem();
-        personselected.setPulsaciones(Integer.parseInt(edittedCell.getNewValue().toString()));
+        personselected.setPulsaciones(Float.parseFloat(edittedCell.getNewValue().toString()));
     }
